pyproject.toml
[tool.poetry]
name = "attestation"
version = "0.1.0"
description = ""
authors = ["BlknDm <89090018300@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
envparse = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
